requirements.txt
discord.py
ffmpeg-python
youtube-dl
python-dotenv